Fix Constant conversions, part 12

+++ main4370.cpp
@@ -337,7 +337,7 @@ case 'b':
 W5y5eD = f3WP((char)-1760044043, (unsigned char)1748142654, (unsigned char)284940313, U_5_jcD48n, (unsigned short)-135883065);
 break;
 case 'c':
-grOyf41 = f3WP((char)-183645296, (unsigned char)-210826812, (unsigned char)-837035397, 133184962, -1640337997);
+grOyf41 = f3WP((char)-183645296, (unsigned char)-210826812, (unsigned char)-837035397, (short)133184962, -1640337997);
 break;
 case 'd':
 GGeT_h78 = Q34482_DqB - L__ic627;
