+++ main663.cpp
@@ -39,7 +39,7 @@ std::cout << HW5V((unsigned short)-1926648370, TM1V_, u8mGeEtYiQ, (short)-565660
 
 std::cout << HW5V((unsigned short)20125775, (unsigned char)1947883060, -1833778384, (short)-1779450949, m0FAIU_q) / HW5V(yg217q, (unsigned char)396673675, u8mGeEtYiQ, fFF5X8lz2, (short)823297453) << std::endl;
 
-std::cout << "tLm90g5C_" << "bzLl_rzN6l" << HW5V((unsigned short)928242305, (unsigned char)-624543988, u8mGeEtYiQ, fFF5X8lz2, 358039075) << "A7Z5uBR72" << std::endl;
+std::cout << "tLm90g5C_" << "bzLl_rzN6l" << HW5V((unsigned short)928242305, (unsigned char)-624543988, u8mGeEtYiQ, fFF5X8lz2, (short)358039075) << "A7Z5uBR72" << std::endl;
 if (U54sMx97We <= -136926476)
 {
 fFF5X8lz2 = (double)(rr2M6a75G);
